use dynamic import to import global components
